Criei o alteraInfluencia, Satisfacao e o enum Influencia

+++ hva-skeleton/hva/core/Satisfacao.java
@@ -1,5 +1,5 @@
 package hva.core;
 
 public interface Satisfacao {
-  public double satisfacao();
+  public int satisfacao();
 }
